# Performance regression for events without tags