References #585 

`akka.persistence.jdbc.integration.PostgresJournalPerfSpec` on master against local docker postgres:
```
[info] A PersistentActor's performance
[info] - must measure: persist()-ing 100 events for 100 actors (10 seconds, 218 milliseconds)
[info]   + Persist()-ing 100 * 100 took 10184 ms
[info]   + Average time: 10184 ms
[info] A PersistentActor's performance
[info] - must measure: persistAsync()-ing 100 events for 100 actors (3 seconds, 984 milliseconds)
[info]   + persistAsync()-ing 100 * 100 took 3955 ms
[info]   + Average time: 3955 ms
```

With this fix, back to pre-5.0.0 levels:
```
[info] A PersistentActor's performance
[info] - must measure: persist()-ing 100 events for 100 actors (2 seconds, 244 milliseconds)
[info]   + Persist()-ing 100 * 100 took 2213 ms
[info]   + Average time: 2213 ms
[info] A PersistentActor's performance
[info] - must measure: persistAsync()-ing 100 events for 100 actors (441 milliseconds)
[info]   + persistAsync()-ing 100 * 100 took 408 ms
[info]   + Average time: 408 ms
```

Note that tagged events (and batches with tagged events in them) will still be in the slow branch of writing.